extract multi_cluster by interface
